requirements.txt
# Python Dependencies for EFMM-Toll System
# Core dependencies for all modules

# Async framework
asyncio-mqtt==0.16.1
aiohttp==3.9.1
aiofiles==23.2.1

# Machine Learning
torch==2.1.2
torchvision==0.16.2
numpy==1.26.2
scikit-learn==1.3.2

# Data handling
pandas==2.1.4
pyyaml==6.0.1

# Cryptography (placeholder - would need actual crypto libraries)
cryptography==41.0.8
pynacl==1.5.0

# Database
aiosqlite==0.19.0

# Configuration
python-dotenv==1.0.0
click==8.1.7

# Logging and monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development
black==23.12.1
flake8==6.1.0
mypy==1.8.0

# Optional V2X/Automotive (would need specialized libraries)
# canopen==2.2.0  # CAN bus communication
# j1939==1.0.2    # J1939 protocol

# Web interface (if needed)
fastapi==0.104.1
uvicorn==0.24.0

# Utilities
python-dateutil==2.8.2
requests==2.31.0
